refactor(CoreS3 SE): replace custom components with M5Stack official
axp2101, aw88298, aw9523b moved to archived/. All three had subtle
incompatibilities with each other (aw9523b mode_mask_ init, pin schema
key difference, aw88298 setup_priority magic number) that caused OTA
rollback crashes. M5Stack official components are vendor-validated and
crash-free.

m5cores3se.yaml rewritten: full VA state machine, media_player
announcement pipeline, 3 wake words + stop + VAD, mic mute,
engine/sensitivity selects, 9-page display, bluetooth_proxy.

Also: aw9523 dump_config format string fix (%.0f → %u).

See DECISIONS.md for component selection framework going forward.

+++ components/aw9523/aw9523.cpp
@@ -65,7 +65,7 @@ void AW9523Component::dump_config() {
   if (this->is_failed()) {
     ESP_LOGE(TAG, "Communication with AW9523 failed!");
   }
-  ESP_LOGCONFIG(TAG, "  Update Interval: %.0f ms", this->get_update_interval());
+  ESP_LOGCONFIG(TAG, "  Update Interval: %u ms", (unsigned) this->get_update_interval());
   // 如果需要，可透過迭代已註冊的 pin 來記錄 pin 的設定
 }
